ef-ptp: Remove the padding field
Remove the padding field from the ptp-sync frames. The reason is that
when we calculate the message length field in the ptp frame, we add also
the padding field which is not really a field of the ptp frame. So then
we get the wrong message length. The fix consists in removing this
padding field. Also it looks like already ef make sure to pad the frames
to be 60 bytes when transmitted.